move tests of MDnsObservatoryDiscovery to its own file
